+++ action.ts
@@ -1,4 +1,112 @@
-import { QueryState } from "./types.ts";
+import {
+  call,
+  createContext,
+  each,
+  Operation,
+  Signal,
+  SignalQueueFactory,
+  spawn,
+  Stream,
+} from "./deps.ts";
+import { ActionPattern, matcher } from "./matcher.ts";
+import type { ActionWPayload, AnyAction } from "./types.ts";
+import { createFilterQueue } from "./queue.ts";
+import { Action } from "./query/types.ts";
+
+export const ActionContext = createContext<Signal<AnyAction, void>>(
+  "store:action",
+);
+
+export function useActions(pattern: ActionPattern): Stream<AnyAction, void> {
+  return {
+    *subscribe() {
+      const actions = yield* ActionContext;
+      const match = matcher(pattern);
+      yield* SignalQueueFactory.set(() => createFilterQueue(match) as any);
+      return yield* actions.subscribe();
+    },
+  };
+}
+
+export function emit({
+  signal,
+  action,
+}: {
+  signal: Signal<AnyAction, void>;
+  action: AnyAction | AnyAction[];
+}) {
+  if (Array.isArray(action)) {
+    if (action.length === 0) {
+      return;
+    }
+    action.map((a) => signal.send(a));
+  } else {
+    signal.send(action);
+  }
+}
+
+export function* put(action: AnyAction | AnyAction[]) {
+  const signal = yield* ActionContext;
+  return emit({
+    signal,
+    action,
+  });
+}
+
+export function take<P>(pattern: ActionPattern): Operation<ActionWPayload<P>>;
+export function* take(pattern: ActionPattern): Operation<Action> {
+  const fd = useActions(pattern);
+  for (const action of yield* each(fd)) {
+    return action;
+  }
+
+  return { type: "take failed, this should not be possible" };
+}
+
+export function* takeEvery<T>(
+  pattern: ActionPattern,
+  op: (action: Action) => Operation<T>,
+) {
+  return yield* spawn(function* (): Operation<void> {
+    const fd = useActions(pattern);
+    for (const action of yield* each(fd)) {
+      yield* spawn(() => op(action));
+      yield* each.next();
+    }
+  });
+}
+
+export function* takeLatest<T>(
+  pattern: ActionPattern,
+  op: (action: Action) => Operation<T>,
+) {
+  return yield* spawn(function* (): Operation<void> {
+    const fd = useActions(pattern);
+    let lastTask;
+
+    for (const action of yield* each(fd)) {
+      if (lastTask) {
+        yield* lastTask.halt();
+      }
+      lastTask = yield* spawn(() => op(action));
+      yield* each.next();
+    }
+  });
+}
+export const latest = takeLatest;
+
+export function* takeLeading<T>(
+  pattern: ActionPattern,
+  op: (action: Action) => Operation<T>,
+) {
+  return yield* spawn(function* (): Operation<void> {
+    while (true) {
+      const action = yield* take(pattern);
+      yield* call(() => op(action));
+    }
+  });
+}
+export const leading = takeLeading;
 
 export const API_ACTION_PREFIX = "@@starfx";
 export const createAction = (curType: string) => {
@@ -7,12 +115,4 @@ export const createAction = (curType: string) => {
   const action = () => ({ type });
   action.toString = () => type;
   return action;
-};
-
-export const createQueryState = (s: Partial<QueryState> = {}): QueryState => {
-  return {
-    "@@starfx/loaders": {},
-    "@@starfx/data": {},
-    ...s,
-  };
 };
