+++ query/thunk.test.ts
@@ -1,8 +1,7 @@
 import { assertLike, asserts, describe, it } from "../test.ts";
-import { configureStore, put, takeEvery } from "../store/mod.ts";
+import { configureStore } from "../store/mod.ts";
+import { put, takeEvery } from "../action.ts";
 import { call, sleep as delay } from "../deps.ts";
-import type { QueryState } from "../types.ts";
-import { createQueryState } from "../action.ts";
 import { sleep } from "../test.ts";
 import { createThunks } from "./thunk.ts";
 import type { Next, ThunkCtx } from "./types.ts";
@@ -52,7 +51,7 @@ const deserializeTicket = (u: TicketResponse): Ticket => {
   };
 };
 
-interface TestState extends QueryState {
+interface TestState {
   users: { [key: string]: User };
   tickets: { [key: string]: Ticket };
 }
@@ -134,14 +133,13 @@ it(
     const fetchUsers = api.create(`/users`, { supervisor: takeEvery });
 
     const store = configureStore<TestState>({
-      initialState: { ...createQueryState(), users: {}, tickets: {} },
+      initialState: { users: {}, tickets: {} },
     });
     store.run(api.bootup);
 
     store.dispatch(fetchUsers());
 
     asserts.assertEquals(store.getState(), {
-      ...createQueryState(),
       users: { [mockUser.id]: deserializeUser(mockUser) },
       tickets: {},
     });
@@ -172,13 +170,12 @@ it(
     });
 
     const store = configureStore<TestState>({
-      initialState: { ...createQueryState(), users: {}, tickets: {} },
+      initialState: { users: {}, tickets: {} },
     });
     store.run(api.bootup);
 
     store.dispatch(fetchTickets());
     asserts.assertEquals(store.getState(), {
-      ...createQueryState(),
       users: { [mockUser.id]: deserializeUser(mockUser) },
       tickets: { [mockTicket.id]: deserializeTicket(mockTicket) },
     });
@@ -256,7 +253,6 @@ it(tests, "error inside endpoint mdw", () => {
 
   const store = configureStore({
     initialState: {
-      ...createQueryState(),
       users: {},
     },
   });
