+++ store/context.ts
@@ -1,11 +1,6 @@
-import { Channel, createChannel, createContext, Signal } from "../deps.ts";
-import type { AnyAction, AnyState } from "../types.ts";
-
+import { Channel, createChannel, createContext } from "../deps.ts";
+import type { AnyState } from "../types.ts";
 import type { FxStore } from "./types.ts";
-
-export const ActionContext = createContext<Signal<AnyAction, void>>(
-  "store:action",
-);
 
 export const StoreUpdateContext = createContext<Channel<void, void>>(
   "store:update",
