+++ store/fx.ts
@@ -1,21 +1,9 @@
-import {
-  call,
-  each,
-  Operation,
-  Result,
-  Signal,
-  SignalQueueFactory,
-  spawn,
-  Stream,
-} from "../deps.ts";
-import { ActionPattern, matcher } from "../matcher.ts";
-import type { ActionWPayload, AnyAction, AnyState } from "../types.ts";
+import { Operation, Result } from "../deps.ts";
+import type { AnyState } from "../types.ts";
 import type { FxStore, StoreUpdater, UpdaterCtx } from "./types.ts";
-import { ActionContext, StoreContext } from "./context.ts";
-import { createFilterQueue } from "../queue.ts";
+import { StoreContext } from "./context.ts";
 import { LoaderOutput } from "./slice/loader.ts";
 import { safe } from "../fx/mod.ts";
-import { Action } from "../query/types.ts";
 
 export function* updateStore<S extends AnyState>(
   updater: StoreUpdater<S> | StoreUpdater<S>[],
@@ -25,23 +13,6 @@ export function* updateStore<S extends AnyState>(
   const st = store as FxStore<S>;
   const ctx = yield* st.update(updater);
   return ctx;
-}
-
-export function emit({
-  signal,
-  action,
-}: {
-  signal: Signal<AnyAction, void>;
-  action: AnyAction | AnyAction[];
-}) {
-  if (Array.isArray(action)) {
-    if (action.length === 0) {
-      return;
-    }
-    action.map((a) => signal.send(a));
-  } else {
-    signal.send(action);
-  }
 }
 
 export function select<S, R>(selectorFn: (s: S) => R): Operation<R>;
@@ -56,80 +27,6 @@ export function* select<S, R, P>(
   const store = yield* StoreContext;
   return selectorFn(store.getState() as S, p);
 }
-
-export function* put(action: AnyAction | AnyAction[]) {
-  const signal = yield* ActionContext;
-  return emit({
-    signal,
-    action,
-  });
-}
-
-function useActions(pattern: ActionPattern): Stream<AnyAction, void> {
-  return {
-    *subscribe() {
-      const actions = yield* ActionContext;
-      const match = matcher(pattern);
-      yield* SignalQueueFactory.set(() => createFilterQueue(match) as any);
-      return yield* actions.subscribe();
-    },
-  };
-}
-
-export function take<P>(pattern: ActionPattern): Operation<ActionWPayload<P>>;
-export function* take(pattern: ActionPattern): Operation<Action> {
-  const fd = useActions(pattern);
-  for (const action of yield* each(fd)) {
-    return action;
-  }
-
-  return { type: "take failed, this should not be possible" };
-}
-
-export function* takeEvery<T>(
-  pattern: ActionPattern,
-  op: (action: Action) => Operation<T>,
-) {
-  return yield* spawn(function* (): Operation<void> {
-    const fd = useActions(pattern);
-    for (const action of yield* each(fd)) {
-      yield* spawn(() => op(action));
-      yield* each.next();
-    }
-  });
-}
-
-export function* takeLatest<T>(
-  pattern: ActionPattern,
-  op: (action: Action) => Operation<T>,
-) {
-  return yield* spawn(function* (): Operation<void> {
-    const fd = useActions(pattern);
-    let lastTask;
-
-    for (const action of yield* each(fd)) {
-      if (lastTask) {
-        yield* lastTask.halt();
-      }
-      lastTask = yield* spawn(() => op(action));
-      yield* each.next();
-    }
-  });
-}
-export const latest = takeLatest;
-
-export function* takeLeading<T>(
-  pattern: ActionPattern,
-  op: (action: Action) => Operation<T>,
-) {
-  return yield* spawn(function* (): Operation<void> {
-    while (true) {
-      const action = yield* take(pattern);
-      yield* call(() => op(action));
-    }
-  });
-}
-export const leading = takeLeading;
 
 export function createTracker<T, M extends Record<string, unknown>>(
   loader: LoaderOutput<M, AnyState>,
