+++ store/store.test.ts
@@ -2,8 +2,9 @@ import { createScope, Operation, Result } from "../deps.ts";
 import { parallel } from "../fx/mod.ts";
 import { asserts, describe, it } from "../test.ts";
 import { StoreContext, StoreUpdateContext } from "./context.ts";
-import { put, take, updateStore } from "./fx.ts";
+import { updateStore } from "./fx.ts";
 import { configureStore, createStore } from "./store.ts";
+import { put, take } from "../action.ts";
 
 const tests = describe("store");
