+++ store/supervisor.ts
@@ -1,5 +1,5 @@
 import { race } from "../fx/mod.ts";
-import { take } from "./fx.ts";
+import { take } from "../action.ts";
 import { call, Callable, Operation, sleep, spawn, Task } from "../deps.ts";
 import type { ActionWPayload, AnyAction } from "../types.ts";
 import type { CreateActionPayload } from "../query/mod.ts";
